etc/qutebrowser: Reorder statusbar widgets, add session name
Move the progress bar to the left of the URL, as it seems more
reasonable to have it close to the URL when loading a page, and it
causes less "jumping" of the widgets in the lower right corner.

Also add the name of the session to the statusbar, so we know which
session we are in (if using multiple sessions/windows on the same
desktop view).